refactor: extract admin memory projection
